==> outlier_test/04_TimeseriesPlots_2019-2021 - Final.R <==
#--------------------------------------------------------------------------------------------------------------#
#--Creates Timeseries Plots to Evaluate AWLN Data
#--Program created by: Erica DiFilippo
#-- Last run by hpham, 02/04/2022
#--------------------------------------------------------------------------------------------------------------#

#--------------------------------------------------------------------------------------------------------------#
#--Set working directory--#
DIR <- 'c:/Users/hpham/OneDrive - INTERA Inc/projects/045_100AreaPT/t100areaPT/'
setwd(DIR)
#--------------------------------------------------------------------------------------------------------------#

#--------------------------------------------------------------------------------------------------------------#
#--Import R Libraries--#
library(data.table)
library(sspaplottools)
#--------------------------------------------------------------------------------------------------------------#

#--------------------------------------------------------------------------------------------------------------#
#--Import Data--#
load('1a_AWLN_Outlier_Analysis/Output/DataPull_020222/DATA_FINAL.RData')
load('0_Data/Well_Data/DataPull_020222/WELL.RData')
#--------------------------------------------------------------------------------------------------------------#

#--------------------------------------------------------------------------------------------------------------#
#--Load User-Defined functions--#
source('1a_AWLN_Outlier_Analysis/RFunctions/pltAWLNTS.R')
#--------------------------------------------------------------------------------------------------------------#

#--------------------------------------------------------------------------------------------------------------#
#--Extract AWLN Wells--#
WELLS <- data.table(NAME=sort(unique(c(DATA$AWLN$NAME))))
#--------------------------------------------------------------------------------------------------------------#

#--------------------------------------------------------------------------------------------------------------#
#--Designate OU for Each Well--#
setkey(WELLS,'NAME')
setkey(AREA,'NAME')
WELLS <- AREA[WELLS]
WELLS <- WELLS[!is.na(OU)]
#--------------------------------------------------------------------------------------------------------------#

#--------------------------------------------------------------------------------------------------------------#
#--Extract OUs--#
OUs <- sort(unique(WELLS$OU))
#--------------------------------------------------------------------------------------------------------------#

#--------------------------------------------------------------------------------------------------------------#
#--Loop through Each OU--#
for (i in 1:length(OUs)){
  
  #--------------------------------------------------------------------------------------------------------------#
  #--Extract Wells by OU--#
  well_sub <- WELLS[OU == OUs[i]]$NAME
  #--------------------------------------------------------------------------------------------------------------#
  
  #--------------------------------------------------------------------------------------------------------------#
  #--Loop through all Analytes--#
  for(n in 1:length(well_sub)){
    
    #--------------------------------------------------------------------------------------------------------------#
    #--Subset dataset by analyte--#
    AWLN_SUB <- subset(DATA$AWLN,DATA$AWLN$NAME==well_sub[n] & year(DATA$AWLN$EVENT) >= 2018)
    MAN_SUB <- subset(DATA$MAN,DATA$MAN$NAME==well_sub[n] & year(DATA$MAN$EVENT) >= 2018)
    ELEVBOT_SUB <- subset(ELEVBOT,NAME==well_sub[n])
    ELEVTOP_SUB <- subset(COORDS,NAME==well_sub[n])
    SCREEN_SUB <- subset(SCREEN,NAME==well_sub[n])
    #--------------------------------------------------------------------------------------------------------------#
    
    #--------------------------------------------------------------------------------------------------------------#
    #--Well Geometry Information--#
    BOT <- ELEVBOT_SUB$ELEVBOT
    TOP <- ELEVTOP_SUB$ZCOORDS
    BOS <- SCREEN_SUB$BOT
    TOS <- SCREEN_SUB$TOP
    
    if(length(BOT) == 0) {BOT = NA}
    if(length(TOP) == 0) {TOP = NA}
    if(length(BOS) == 0) {BOS = NA}
    if(length(TOS) == 0) {TOS = NA}
    #--------------------------------------------------------------------------------------------------------------#
    
    #--------------------------------------------------------------------------------------------------------------#
    #--Designate File Name--#
    fn <- paste0('1a_AWLN_Outlier_Analysis/Output/DataPull_020222/Final Evaluation/',OUs[i],'/',well_sub[n],'.png')
    #--------------------------------------------------------------------------------------------------------------#
    
    #--------------------------------------------------------------------------------------------------------------#
    #--Plot Timeseries--#
    if(nrow(AWLN_SUB) > 0){
      png(fn,height=8.5,width=11,units='in',res=600)
      pltAWLNTS(AWLN_SUB,MAN_SUB,BOT,TOP,BOS,TOS,mindate=ISOdate(2019,01,01),maxdate=ISOdate(2021,12,31),ts='Yearly')
      dev.off()
    }
    #--------------------------------------------------------------------------------------------------------------#
    
  }
  #--------------------------------------------------------------------------------------------------------------#
  
}
#--------------------------------------------------------------------------------------------------------------#
